mcp-client/requirements.txt
fastmcp==2.11.3
uvicorn==0.35.0
pyshark==0.6
scapy==2.6.1
manuf==1.1.5
manuf2==2.0.1
aiohttp==3.12.15
quart
redis
asyncio
nest_asyncio
quart-auth
quart-wtforms
email_validator
quart-wtforms[email]
setuptools
passlib
bcrypt
ansible
ansible-runner
quart-rate-limiter
iperf3
tzlocal
pyjwt
pysnmp
